@@ -1,24 +1,24 @@
 module github.com/pflege-de/connect
 
-go 1.17
-
-require github.com/nats-io/nats.go v1.13.1-0.20220121202836-972a071d373d
+go 1.22.6
 
 require (
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/nats-io/nats-server/v2 v2.7.2 // indirect
-	golang.org/x/net v0.0.0-20211112202133-69e39bad7dc2 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.27.1 // indirect
+	github.com/golang-jwt/jwt/v4 v4.5.0
+	github.com/google/uuid v1.6.0
+	github.com/nats-io/nats.go v1.37.0
+	github.com/pflege-de/go-force v1.1.2
+	github.com/pkg/errors v0.9.1
+	github.com/vmihailenco/msgpack v4.0.4+incompatible
 )
 
 require (
-	github.com/golang-jwt/jwt/v4 v4.2.0
-	github.com/google/uuid v1.3.0
-	github.com/nats-io/nkeys v0.3.0 // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/klauspost/compress v1.17.2 // indirect
+	github.com/nats-io/nkeys v0.4.7 // indirect
 	github.com/nats-io/nuid v1.0.1 // indirect
-	github.com/nimajalali/go-force v0.0.0-20200831220737-454890ee2b7c
-	github.com/pkg/errors v0.9.1
-	github.com/vmihailenco/msgpack v4.0.4+incompatible
-	golang.org/x/crypto v0.0.0-20220112180741-5e0467b6c7ce // indirect
+	golang.org/x/crypto v0.18.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	google.golang.org/appengine v1.6.8 // indirect
+	google.golang.org/protobuf v1.26.0 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
