@@ -2,16 +2,16 @@ package authentication
 
 import (
 	"crypto/x509"
-	"io/ioutil"
+	"io"
 	"log"
 	"net/http"
 	"os"
 	"strings"
 
-	"github.com/nimajalali/go-force/force"
+	"github.com/pflege-de/go-force/force"
 )
 
-func NewForce() (*force.ForceApi, error) {
+func NewForce() (*force.ForceApiSObjectInterface, error) {
 	sfRequest := AuthenticationRequest{
 		URL:      os.Getenv("SF_SCAUD"),
 		Username: os.Getenv("SF_SCUSER"),
@@ -28,14 +28,17 @@ func NewForce() (*force.ForceApi, error) {
 		return nil, err
 	}
 
-	return force.CreateWithAccessToken(
+	fapi, err := force.CreateWithAccessToken(
 		"v53.0",
 		os.Getenv("SF_CLIENT_ID"),
 		authReponse.GetToken(),
-		os.Getenv("SF_SCINSTANCE"))
+		os.Getenv("SF_SCINSTANCE"),
+		http.DefaultClient,
+	)
+	return &fapi, err
 }
 
-func NewForceKeyStringSecret() (*force.ForceApi, error) {
+func NewForceKeyStringSecret() (*force.ForceApiSObjectInterface, error) {
 	sfRequest := AuthenticationRequest{
 		URL:      os.Getenv("SF_SCAUD"),
 		Username: os.Getenv("SF_SCUSER"),
@@ -50,7 +53,7 @@ func NewForceKeyStringSecret() (*force.ForceApi, error) {
 	}
 	log.Println(key)
 
-	r := ioutil.NopCloser(strings.NewReader(string("test"))) // r type is io.ReadCloser
+	r := io.NopCloser(strings.NewReader("test")) // r type is io.ReadCloser
 	defer r.Close()
 
 	authReponse, err := Authenticate(sfRequest, r, http.DefaultClient)
@@ -58,9 +61,12 @@ func NewForceKeyStringSecret() (*force.ForceApi, error) {
 		return nil, err
 	}
 
-	return force.CreateWithAccessToken(
+	fapi, err := force.CreateWithAccessToken(
 		"v53.0",
 		os.Getenv("SF_CLIENT_ID"),
 		authReponse.GetToken(),
-		os.Getenv("SF_SCINSTANCE"))
+		os.Getenv("SF_SCINSTANCE"),
+		http.DefaultClient,
+	)
+	return &fapi, err
 }
