@@ -3,11 +3,11 @@ package connect
 import (
 	"log"
 
-	"github.com/nimajalali/go-force/force"
 	"github.com/pflege-de/connect/authentication"
+	"github.com/pflege-de/go-force/force"
 )
 
-func GetForceApi() *force.ForceApi {
+func GetForceApi() *force.ForceApiSObjectInterface {
 	forceApi, err := authentication.NewForce()
 	if err != nil {
 		log.Printf("Couldn't establish forceApi: %s", err)
@@ -15,7 +15,7 @@ func GetForceApi() *force.ForceApi {
 	return forceApi
 }
 
-func GetForceApiKeyStringSecret() *force.ForceApi {
+func GetForceApiKeyStringSecret() *force.ForceApiSObjectInterface {
 	forceApi, err := authentication.NewForceKeyStringSecret()
 	if err != nil {
 		log.Printf("Couldn't establish forceApi: %s", err)
@@ -23,7 +23,7 @@ func GetForceApiKeyStringSecret() *force.ForceApi {
 	return forceApi
 }
 
-func GetForceApiOAuth() *force.ForceApi {
+func GetForceApiOAuth() *force.ForceApiSObjectInterface {
 	forceApi, err := authentication.NewOAuthForce()
 	if err != nil {
 		log.Printf("Couldn't establish forceApi: %s", err)
